+++ packages/scaffold/src/partials/w3m-account-default-widget/index.ts
@@ -18,6 +18,7 @@ import { LitElement, html } from 'lit'
 import { state } from 'lit/decorators.js'
 import { ifDefined } from 'lit/directives/if-defined.js'
 import styles from './styles.js'
+import { W3mFrameRpcConstants } from '@web3modal/wallet'
 
 @customElement('w3m-account-default-widget')
 export class W3mAccountDefaultWidget extends LitElement {
@@ -120,7 +121,7 @@ export class W3mAccountDefaultWidget extends LitElement {
       </wui-flex>
 
       <wui-flex flexDirection="column" gap="xs" .padding=${['0', 's', 's', 's'] as const}>
-        ${this.emailCardTemplate()} <w3m-account-auth-button></w3m-account-auth-button>
+        ${this.authCardTemplate()} <w3m-account-auth-button></w3m-account-auth-button>
 
         <wui-list-item
           .variant=${networkImage ? 'image' : 'icon'}
@@ -179,7 +180,7 @@ export class W3mAccountDefaultWidget extends LitElement {
     `
   }
 
-  private emailCardTemplate() {
+  private authCardTemplate() {
     const type = StorageUtil.getConnectedConnector()
     const authConnector = ConnectorController.getAuthConnector()
     const { origin } = location
@@ -245,7 +246,15 @@ export class W3mAccountDefaultWidget extends LitElement {
   }
 
   private onTransactions() {
-    EventsController.sendEvent({ type: 'track', event: 'CLICK_TRANSACTIONS' })
+    EventsController.sendEvent({
+      type: 'track',
+      event: 'CLICK_TRANSACTIONS',
+      properties: {
+        isSmartAccount:
+          AccountController.state.preferredAccountType ===
+          W3mFrameRpcConstants.ACCOUNT_TYPES.SMART_ACCOUNT
+      }
+    })
     RouterController.push('Transactions')
   }