+++ examples/react-wagmi/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @examples/react-wagmi
 
+## 5.0.0
+
+### Major Changes
+
+- Release V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/wagmi@5.0.0
+
+## 5.0.0-cn-v5.0
+
+### Major Changes
+
+- Test V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/wagmi@5.0.0-cn-v5.0
+
 ## 4.2.3
 
 ### Patch Changes