+++ packages/scaffold-utils/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @web3modal/scaffold-utils
 
+## 5.0.0
+
+### Major Changes
+
+- Release V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/core@5.0.0
+  - @web3modal/polyfills@5.0.0
+
+## 5.0.0-cn-v5.0
+
+### Major Changes
+
+- Test V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/core@5.0.0-cn-v5.0
+  - @web3modal/polyfills@5.0.0-cn-v5.0
+
 ## 4.2.3
 
 ### Patch Changes